drivers: mpsl: fix TEMP_NRF5_MPSL option
With
https://github.com/zephyrproject-rtos/zephyr/commit/f0441fba58a242d4fcc70794f6250ace358baae4
merged upstream, `TEMP_NRF5` is enabled by default if temperature sensor
is defined in devicetree. That caused build errors, as `TEMP_NRF5_MPSL`
option required `TEMP_NRF5` to be disabled.

This commit fixes that error by treating `TEMP_NRF5_MPSL` as an
alternative implementation of temperature driver, selected through
`TEMP_NRF5_FORCE_ALT`. The intended usage of that symbol is to allow the
application to select `TEMP_NRF5`, but to use a different underlying
source code for actually driving the sensor. `TEMP_NRF5_MPSL` is aligned
to achieve that.

Signed-off-by: Jedrzej Ciupis <jedrzej.ciupis@nordicsemi.no>