sysbuild: introduce a configure_cache
Introduce a configure_cache function which allows to update the CMake
cache file of an image when the cache file contains variable values
of the form:
<VAR>:<TYPE>=...@VAR_NAME_TO_UPDATE@...

This allows an image to store a CMake variable in the cache and have
sysbuild replace such var with correct value during later sysbuild
CMake stage.

As practical example, this allows anm image to specify a partition
manager variable, such as PM_<image>_SIZE, and continue.
When partition manager has completed, then the variable will be updated
with correct value.

Signed-off-by: Torsten Rasmussen <Torsten.Rasmussen@nordicsemi.no>