drivers: wifi: Add support to flash the patch automatically
If the storing of nRF70 FW patch in external flash (non-XIP) is enabled,
user needs to two steps, flash the patch to external flash and regular
Zephyr image flash.

To automate this process:

0. Convert the patch binary to hex using "intelhex" python package.
1. Generate a merged.hex file similar to multi-image build, but combines
application image with nRF70 patch image .
2. Set the merged image as default hex file for the flash

This way, without any user involvement we can automate the process.

Signed-off-by: Chaitanya Tata <Chaitanya.Tata@nordicsemi.no>