tests: bl_validation_neg: Fix configuration
The test requries to leave the 1st slot empty as it copies the FW from
the test case body.
It is requried to flash 3 DEBUG keys to match regular expressions.
The last "\n" character is not printed out, which leads to the lack of
"Rebooting..." messages in the test log.

Signed-off-by: Tomasz Chyrowicz <tomasz.chyrowicz@nordicsemi.no>
Signed-off-by: Jamie McCrae <jamie.mccrae@nordicsemi.no>